+++ examples/src/js/ClickableLabelsExample.js
@@ -28,7 +28,7 @@ const nodes = [
                 value: '/app/Providers',
                 label: 'Providers',
                 children: [{
-                    value: '/app/Http/Providers/EventServiceProvider.js',
+                    value: '/app/Providers/EventServiceProvider.js',
                     label: 'EventServiceProvider.js',
                 }],
             },
@@ -125,6 +125,7 @@ class ClickExample extends React.Component {
                 <CheckboxTree
                     checked={checked}
                     expanded={expanded}
+                    iconsClass="fa5"
                     nodes={nodes}
                     expandOnClick
                     onCheck={this.onCheck}