+++ gulpfile.js
@@ -7,8 +7,8 @@ const gulp = require('gulp');
 const header = require('gulp-header');
 const less = require('gulp-less');
 const mocha = require('gulp-mocha');
-const sass = require('gulp-sass');
-const scsslint = require('gulp-scss-lint');
+const sass = require('gulp-dart-sass');
+const styleLint = require('gulp-stylelint');
 const webpack = require('webpack');
 const webpackStream = require('webpack-stream');
 
@@ -41,7 +41,7 @@ gulp.task('test-script-mocha', () => (
         }))
 ));
 
-gulp.task('test-script', gulp.series('test-script-format', 'test-script-mocha'));
+gulp.task('test-script', gulp.series(gulp.parallel('test-script-format', 'test-script-mocha')));
 
 gulp.task('build-script', gulp.series('test-script', () => (
     gulp.src(['./src/index.js'])
@@ -59,14 +59,15 @@ gulp.task('build-script-web', gulp.series('build-script', () => (
 
 gulp.task('build-style', () => (
     gulp.src('./src/scss/**/*.scss')
-        .pipe(scsslint())
-        .pipe(scsslint.failReporter())
+        .pipe(styleLint({
+            reporters: [
+                { formatter: 'string', console: true },
+            ],
+        }))
         .pipe(sass({
             outputStyle: 'expanded',
         }).on('error', sass.logError))
-        .pipe(autoprefixer({
-            browsers: ['last 2 versions'],
-        }))
+        .pipe(autoprefixer())
         .pipe(gulp.dest('./lib'))
         .pipe(cleanCss())
         .pipe(gulp.dest('./.css-compare/scss'))
@@ -75,9 +76,7 @@ gulp.task('build-style', () => (
 gulp.task('build-style-less', () => (
     gulp.src('./src/less/**/*.less')
         .pipe(less())
-        .pipe(autoprefixer({
-            browsers: ['last 2 versions'],
-        }))
+        .pipe(autoprefixer())
         .pipe(cleanCss())
         .pipe(gulp.dest('./.css-compare/less'))
 ));
@@ -98,14 +97,15 @@ function buildExamplesScript(mode = 'development') {
 
 function buildExamplesStyle(minifyStyles = false) {
     let stream = gulp.src('./examples/src/scss/**/*.scss')
-        .pipe(scsslint())
-        .pipe(scsslint.failReporter())
+        .pipe(styleLint({
+            reporters: [
+                { formatter: 'string', console: true },
+            ],
+        }))
         .pipe(sass({
             outputStyle: 'expanded',
         }).on('error', sass.logError))
-        .pipe(autoprefixer({
-            browsers: ['last 2 versions'],
-        }));
+        .pipe(autoprefixer());
 
     if (minifyStyles) {
         stream = stream.pipe(cleanCss());
@@ -132,7 +132,7 @@ gulp.task('build-examples-style-prod', () => (
 
 gulp.task('build-examples-html', () => (
     gulp.src('./examples/src/index.html')
-        .pipe(gulp.dest('./examples/dist/'))
+        .pipe(gulp.dest('./examples/dist'))
         .pipe(browserSync.stream())
 ));